fix(home): correct prefetch implementation
- use tanstack hydration boundary
  - common queryClient util
- move listCampaigns (/api/campaigns::GET) into @/lib/api/campaigns
