+++ app/api/campaigns/route.ts
@@ -8,14 +8,9 @@ import {
 } from '@/lib/api/error';
 import { response, handleError } from '@/lib/api/response';
 
-import { createPublicClient, http } from 'viem';
-import { celoAlfajores } from 'viem/chains';
-import { DbCampaign } from '@/types/campaign';
-import { chainConfig } from '@/lib/web3/config/chain';
 import { CampaignStatus } from '@/types/campaign';
+import { listCampaigns } from '@/lib/api/campaigns';
 
-const FACTORY_ADDRESS = process.env.NEXT_PUBLIC_CAMPAIGN_INFO_FACTORY;
-const RPC_URL = chainConfig.rpcUrl;
 const statusMap: Record<string, CampaignStatus> = {
   draft: CampaignStatus.DRAFT,
   pending_approval: CampaignStatus.PENDING_APPROVAL,
@@ -70,99 +65,6 @@ async function uploadToCloudinary(file: File): Promise<string> {
   const data = await response.json();
   console.log('Cloudinary upload successful:', data);
   return data.secure_url;
-}
-
-async function getPublicClient() {
-  if (!FACTORY_ADDRESS || !RPC_URL) {
-    throw new Error('Campaign factory address or RPC URL not configured');
-  }
-
-  return createPublicClient({
-    chain: {
-      ...celoAlfajores,
-      contracts: {
-        multicall3: {
-          address: '0xcA11bde05977b3631167028862bE2a173976CA11' as const,
-          blockCreated: 14353601,
-        },
-      },
-    },
-    transport: http(RPC_URL),
-    batch: {
-      multicall: {
-        batchSize: 1024,
-        wait: 16,
-      },
-    },
-  });
-}
-
-async function getCampaignCreatedEvents(
-  client: ReturnType<typeof createPublicClient>,
-) {
-  return client.getLogs({
-    address: FACTORY_ADDRESS as `0x${string}`,
-    event: {
-      type: 'event',
-      name: 'CampaignInfoFactoryCampaignCreated',
-      inputs: [
-        { type: 'bytes32', name: 'identifierHash', indexed: true },
-        { type: 'address', name: 'campaignInfoAddress', indexed: true },
-      ],
-    },
-    fromBlock: 0n,
-    toBlock: 'latest',
-  });
-}
-
-type CampaignCreatedEvent = {
-  args: {
-    identifierHash?: `0x${string}`;
-    campaignInfoAddress?: `0x${string}`;
-  };
-};
-
-function formatCampaignData(
-  dbCampaign: DbCampaign,
-  event: CampaignCreatedEvent | undefined,
-  forceEvent = true,
-) {
-  if (forceEvent && (!event || !event.args)) {
-    console.error('No matching event found for campaign:', {
-      campaignId: dbCampaign.id,
-      campaignAddress: dbCampaign.campaignAddress,
-    });
-    return null;
-  }
-
-  return {
-    id: dbCampaign.id,
-    title: dbCampaign.title,
-    description: dbCampaign.description,
-    status: dbCampaign.status,
-    address: dbCampaign.campaignAddress,
-    owner: dbCampaign.creatorAddress,
-    launchTime: Math.floor(
-      new Date(dbCampaign.startTime).getTime() / 1000,
-    ).toString(),
-    deadline: Math.floor(
-      new Date(dbCampaign.endTime).getTime() / 1000,
-    ).toString(),
-    goalAmount: dbCampaign.fundingGoal,
-    totalRaised:
-      dbCampaign.payments?.reduce((accumulator, payment) => {
-        const value = Number(payment.amount);
-        if (isNaN(value)) {
-          return accumulator;
-        }
-        return accumulator + value;
-      }, 0) ?? 0,
-    images: dbCampaign.images,
-    slug: dbCampaign.slug,
-    location: dbCampaign.location,
-    category: dbCampaign.category,
-    treasuryAddress: dbCampaign.treasuryAddress,
-  };
 }
 
 export async function POST(req: Request) {
@@ -319,7 +221,8 @@ export async function GET(req: Request) {
     const status = searchParams.get('status') || 'active';
     const page = parseInt(searchParams.get('page') || '1');
     const pageSize = parseInt(searchParams.get('pageSize') || '10');
-    const rounds = searchParams.get('rounds') || 'false';
+    const rounds =
+      (searchParams.get('rounds') || 'false') === 'true' ? true : false;
     const syncChain = searchParams.get('sync-chain') || 'false';
     const skip = (page - 1) * pageSize;
     if (pageSize > 10) {
@@ -328,98 +231,17 @@ export async function GET(req: Request) {
     const forceEvents = syncChain === 'true' ? true : false;
     // status active should be enforced if access-token is not admin
     const admin = await isAdmin();
-    const statusList = !admin
-      ? [CampaignStatus.ACTIVE]
-      : status === 'active'
-        ? [CampaignStatus.ACTIVE]
-        : status === 'all'
-          ? [
-              CampaignStatus.DRAFT,
-              CampaignStatus.PENDING_APPROVAL,
-              CampaignStatus.COMPLETED,
-              CampaignStatus.ACTIVE,
-            ]
-          : [
-              CampaignStatus.PENDING_APPROVAL,
-              CampaignStatus.COMPLETED,
-              CampaignStatus.ACTIVE,
-            ];
-    const [dbCampaigns, totalCount] = await Promise.all([
-      db.campaign.findMany({
-        where: {
-          status: {
-            in: statusList,
-          },
-        },
-        include: {
-          images: true,
-          RoundCampaigns: {
-            include: {
-              Round: true,
-            },
-          },
-          payments: {
-            // requires amount to be numeric: _sum: { amount: true },
-            where: { token: 'USDC', type: 'BUY', status: 'confirmed' },
-          },
-        },
-        skip,
-        take: pageSize,
-        orderBy: {
-          createdAt: 'desc',
-        },
-      }),
-      db.campaign.count({
-        where: {
-          status: {
-            in: statusList,
-          },
-        },
-      }),
-    ]);
-    let events: CampaignCreatedEvent[] = [];
-    if (forceEvents) {
-      const client = await getPublicClient();
-      // @ts-expect-error - client issue
-      events = await getCampaignCreatedEvents(client);
-    }
-    const combinedCampaigns = dbCampaigns
-      .filter((campaign) => campaign.transactionHash)
-      .map((dbCampaign) => {
-        const event = events.find(
-          (onChainCampaign) =>
-            onChainCampaign.args?.campaignInfoAddress?.toLowerCase() ===
-            dbCampaign.campaignAddress?.toLowerCase(),
-        ) as CampaignCreatedEvent | undefined;
-        if (rounds === 'true') {
-          return formatCampaignData(
-            {
-              ...dbCampaign,
-              status: dbCampaign.status,
-              rounds:
-                dbCampaign.RoundCampaigns?.map(({ Round }) => ({
-                  id: Round.id,
-                  title: Round.title,
-                })) ?? [],
-            },
-            event,
-            forceEvents,
-          );
-        }
-        return formatCampaignData(dbCampaign, event, forceEvents);
-      })
-      .filter(Boolean);
-
-    return response({
-      campaigns: combinedCampaigns,
-      pagination: {
-        currentPage: page,
+    return response(
+      await listCampaigns({
+        admin,
+        status,
+        page,
         pageSize,
-        totalPages: Math.ceil(totalCount / pageSize),
-        totalItems: totalCount,
-        hasMore: skip + pageSize < totalCount,
-      },
-    });
+        rounds,
+        skip,
+        forceEvents,
+      }),
+    );
   } catch (error: unknown) {
     return handleError(error);
   }
