+++ app/layout.tsx
@@ -5,8 +5,6 @@ import Providers from './providers';
 import { Toaster } from '@/components/ui/toaster';
 import { NetworkCheck } from '@/components/network-check';
 import { PageMainLayout } from '@/components/page/main-layout';
-import { prefetchCampaigns } from '@/lib/hooks/useCampaigns';
-import { QueryClient } from '@tanstack/react-query';
 import { EnvironmentBadge } from '@/components/environment-badge';
 
 const geistSans = localFont({
@@ -24,10 +22,6 @@ export const metadata: Metadata = {
   title: 'Akashic',
   description: 'Fundraising platform for open source projects',
 };
-
-// Prefetch campaigns data
-const queryClient = new QueryClient();
-prefetchCampaigns(queryClient);
 
 export default function RootLayout({
   children,
