+++ app/providers.tsx
@@ -13,17 +13,9 @@ import {
   EnvironmentProvider,
   CollectionProvider,
 } from '@/contexts';
+import { getQueryClient } from '@/lib/query-client';
 
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      staleTime: 60 * 1000, // 1 minute
-      gcTime: 5 * 60 * 1000, // 5 minutes
-      refetchOnWindowFocus: false,
-      retry: 1,
-    },
-  },
-});
+const queryClient = getQueryClient();
 
 export default function Providers({ children }: { children: ReactNode }) {
   return (
